% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annotateByLocation}
\alias{annotateByLocation}
\title{Fast Gene Annotation by Chromosomal Location}
\usage{
annotateByLocation(x, txDb, txAnnotRange = c(1500, 2000), orgDb,
  orgDb_Columns)
}
\arguments{
\item{x}{A \code{GRanges} object or \code{data.frame}, giving the coordinates to annotate.}

\item{txDb}{A \code{TranscriptDb} object, giving the genomic references for the alignement. If \code{txDb} is missing, the default \code{TxDb.Hsapiens.UCSC.hg19.knownGene} will be used.}

\item{txAnnotRange}{A integer vector, giving the window size for the genes' promotor and extended sites. Default is \code{txAnnotRange= c(1500,2000)}, See details.}

\item{orgDb}{A \code{OrganismDb} object, giving the details of the genomic references. If \code{orgDb} is missing, the default \code{org.Hs.eg.db} will be used.}

\item{orgDb_Columns}{A character vector (optional), giving which columns to extract from the \code{orgDb}. Note that if \code{orgDb_Columns} is used, then the user selected columns will be selected instead of the default request on \code{org.Hs.eg.db}, see details section.}
}
\description{
This method allows fast gene annotation from chromosomal location
}

