% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annotateGetInfoFromOrganismtDb}
\alias{annotateGetInfoFromOrganismtDb}
\title{Get gene information using EntrezID and OrganismDb}
\usage{
annotateGetInfoFromOrganismtDb(orgDb, orgDb_Columns, sep_intra,
  verbose = FALSE)
}
\arguments{
\item{orgDb}{A \code{OrganismDb} object. If \code{orgDb} is missing, the default \code{org.Hs.eg.db} will be used.}

\item{orgDb_Columns}{A character vector, giving which columns to extract from the \code{orgDb}. Note that if \code{orgDb_Columns} is used, then the user selected columns will be selected instead of the default request on \code{org.Hs.eg.db}.}

\item{sep_intra}{A character vector, giving the separator character for gene information, see details.}

\item{verbose}{A logical value, indicating if messages should be printed. Default is \code{FALSE}.}
}
\value{
\itemize{
\item txInfo A \code{data.frame} containing the genes' information.
\item orgDb_Key A character vector, giving the EntrezID's column name.
\item orgDb_Columns A character vector, giving the selected column names (see Arguments section).
}
}
\description{
Gets the genes information in an OrganismDb using EntrezID keys.
}
\author{
Alexandre Thiery
}
\keyword{internal}

