% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annotateGetRefFromTranscriptDb}
\alias{annotateGetRefFromTranscriptDb}
\title{Get mapping between reference and gene using TranscriptDb}
\usage{
annotateGetRefFromTranscriptDb(txDb, exons, verbose = FALSE)
}
\arguments{
\item{txDb}{A \code{TranscriptDb} object.}

\item{exons}{A logical value, indicating if exons should be extracted.}

\item{verbose}{A logical value, indicating if messages should be printed. Default is \code{FALSE}.}
}
\value{
\itemize{
\item df_GENETR A \code{data.frame} containing the genomic ranges of the transcripts, ordered by chromosome and strand;
\item df_EXON A \code{data.frame} containing the genomic ranges of the exons, ordered by chromosome and strand.
\item stack_EXON A \code{data.frame} containing mappings between exons in \code{df_EXON} and transcripts in \code{df_GENETR}.
\item txDb_Key A character vector, giving the EntrezID's column name.
\item txGroup A \code{data.frame} containing the index of each chromosome and strand pair \code{df_GENETR}. Used to speed up computation
}
}
\description{
Gets the mapping between reference (usually transcripts) and gene's EntrezID using a TranscriptDb object.
}
\author{
Alexandre Thiery
}
\keyword{internal}

