% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{ProbeAnnotator-package}
\alias{ProbeAnnotator}
\alias{ProbeAnnotator-package}
\title{ProbeAnnotator package}
\description{
This package provides fast and user friendly wrappers to C++ functions for gene annotation.
}
\author{
Alexandre Thiery, Pierre Bady
}

