% Generated by roxygen2 (4.0.2): do not edit by hand
\name{searchGene}
\alias{searchGene}
\title{Search a "gene" contained in a character vector}
\usage{
searchGene(x, gene, sep = ";", type = "logical")
}
\arguments{
\item{x}{A character vector.}

\item{gene}{A character vector, the gene (character pattern) of interest.}

\item{sep}{A character vector, the separator between genes (by default, \code{sep=";"}).}

\item{type}{A character indicating the type of values returned by the function. By default the function return a logical vector (\code{type="logical"}). The type \code{"numeric"} and \code{"value"} return position and value respectively.}
}
\value{
The functiuon returns a vector of logical (\code{type="logical"}), numeric (\code{type="numeric"}) or character (\code{type="value"}) depending of the selected type.
}
\description{
This function offers the possibility to search genes (character pattern) contained in a character vector separted by a given separator (semi-colon by default).
}
\examples{
{
genelist <-c("MGMT","MGMT;MGMT","MGMT;EGFR","HOXA1;EGFR;HOXA2","EGFR;MGMT",NA)
searchGene(genelist,"EGFR",sep=";",type="logical")
searchGene(genelist,"EGFR",sep=";",type="numeric")
searchGene(genelist,"EGFR",sep=";",type="value")
}
}
\author{
Pierre Bady
}

