% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transform_TAB}
\alias{transform_TAB}
\title{transform_TAB

A function modify platforms' TAB files to FASTA}
\usage{
transform_TAB(fasta.file, file.out, probe_column, seq_column, hasProbeset)
}
\arguments{
\item{fasta.file}{a character vector giving the name of the input TAB file.}

\item{file.out}{a character vector giving the name of the ouput FASTA file.}

\item{probe_column}{a character vector giving the names of the probe column to extract.}

\item{seq_column}{a character vector giving the names of the sequence column to extract.}

\item{hasProbeset}{a logical value indicating if the platform is arranged in probesets or not.}
}
\description{
transform_TAB

A function modify platforms' TAB files to FASTA
}
\details{
The aim of function step is to transform the platforms sequence TAB files in fasta files format. For example, taking a sequence name from Mouse:
}

