% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bowtie.get.makeAlignmentCommand}
\alias{bowtie.get.makeAlignmentCommand}
\title{bowtie.get.makeAlignmentCommand

A function to create the command to perform alignement with Bowtie.}
\usage{
bowtie.get.makeAlignmentCommand(file.in, file.out, index, bowtie.path,
  index.dir = NULL, p = 0)
}
\arguments{
\item{file.in}{a character vector giving the name of the input FASTA file.}

\item{file.out}{a character vector giving the name of the output BOWTIE map file.}

\item{index}{a character vector giving the name of BOWTIE index to align with.}

\item{bowtie.path}{a character vector giving the path of BOWTIE source files.}

\item{index.dir}{a character vector giving the path of BOWTIE indexes files. Default value is \code{bowtie.path + "/indexes/"}}

\item{p}{a numerical value giving the number of core used for multithreading. If \code{p=0}, then multithreading is not used. Default is code{p=0}.}
}
\description{
bowtie.get.makeAlignmentCommand

A function to create the command to perform alignement with Bowtie.
}

