% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bowtie.makeAlignment}
\alias{bowtie.makeAlignment}
\title{bowtie.makeAlignment

A function to create and execute bowtie alignment.}
\usage{
bowtie.makeAlignment(file.in, file.out, index, bowtie.path, index.dir = NULL,
  p = 0)
}
\arguments{
\item{file.in}{a character vector giving the name of the input FASTA file.}

\item{file.out}{a character vector giving the name of the output BOWTIE map file.}

\item{index}{a character vector giving the name of BOWTIE index to align with.}

\item{bowtie.path}{a character vector giving the path of BOWTIE source files.}

\item{index.dir}{a character vector giving the path of BOWTIE indexes files. Default value is \code{bowtie.path + "/indexes/"}.}

\item{p}{a numerical value giving the number of core used for multithreading. If \code{p=0}, then multithreading is not used. Default is code{p=0}.}
}
\description{
bowtie.makeAlignment

A function to create and execute bowtie alignment.
}

